* Merge from branches/tvo3:   * Remove the xyz component of float3, use overloaded operator[] instead.   * long->int to fix 64 bit compatibility
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1447 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Game/UI/GUI/GUIconsole.h
@@ -26,9 +26,9 @@ protected:
 
 	struct ConsoleLine
 	{
-		ConsoleLine(std::string text1, long time1): time(time1), text(text1)
+		ConsoleLine(std::string text1, int time1): time(time1), text(text1)
 		{}
-		long time;
+		int time;
 		std::string text;
 	};	
 	std::deque<ConsoleLine> lines;
